pyproject.toml
[project]
name = "lineage"
version = "0.1.0"
description = "Python CLI tool to create a lineage graph."
authors = [
    {name = "Ramses Kools",email = "14346781+RamsesKools@users.noreply.github.com"}
]
readme = "README.md"

[tool.poetry]
packages = [{include = "lineage", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<4"
click = ">=8.3.0,<9.0.0"
networkx = "^3.5"
sqlglot = "^27.29.0"
ruamel-yaml = "^0.18.16"
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[tool.poetry.scripts]
lineage = "lineage.cli:cli"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
